passed useState into app
